Passando dados(notificações) através do container
