+++ app/src/components/containers/Atividades/index.js
@@ -1,0 +1,3 @@
+import AtividadesContainer from "./AtividadesContainer";
+
+export default AtividadesContainer;
